change file struct
