+++ package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "type": "module",
   "author": "Kiran Boyle",
   "license": "ISC",
   "dependencies": {
