+++ src/__tests__/index.test.ts
@@ -6,17 +6,28 @@ test("Are Correct Elements Bold", async () => {
   await page.goto("https://www.theguardian.com");
   await page.addScriptTag({ path: "./dist/content.js" });
 
-  const tags = await page.$$eval(
-    "h2, side, span, h3, h4, h5, h6, p",
-    (elements: any) =>
-      elements.map((element) => {
-        return element.innerHTML;
-      })
+  const tags = await page.$$eval("h2, h3, h4, h5, h6", (elements: any) =>
+    elements.map((element) => {
+      return element.innerHTML;
+    })
   );
+
   console.log(tags);
 
-  // console.log(headings);
-  // describe("Test each tag being correctly bolded", () => {});
+  const boldTag = "<b>";
+  const closingBoldTag = "</b>";
+
+  tags.forEach((tag: any) => {
+    const words = tag.split(" ");
+    words.forEach((word: string) => {
+      if (word.length > 3) {
+        const firstTag = word.slice(0, 3);
+        const secondTag = word.slice(6, word.length);
+        expect(firstTag).toEqual(boldTag);
+        expect(secondTag).toEqual(closingBoldTag);
+      }
+    });
+  });
 
   browser.close();
 });
