+++ src/content.ts
@@ -1,3 +1,4 @@
+import { boldWords } from "./utils/index";
 const elements: NodeListOf<HTMLElement> = document.querySelectorAll("body *");
 
 function isBold(element: HTMLElement): boolean {
@@ -10,14 +11,6 @@ function isBold(element: HTMLElement): boolean {
 }
 
 const tags = ["P", "SIDE", "SPAN", "H3", "H4", "H5", "H6"];
-
-function boldWords(sliceLength: number, word: string): string {
-  let slicedChars: string = word.slice(0, sliceLength);
-  const restWord: string = word.slice(sliceLength);
-  let boldChars: string = `<b style="font-weight: 600;">${slicedChars}</b>`;
-  const result: string = boldChars + restWord;
-  return result;
-}
 
 for (let element of elements) {
   if (element.innerText && tags.includes(element.tagName) && !isBold(element)) {
